Fixed illum mode + Basic Transparency in Blender
Since Unity pulls an OBJ as a big file instead of submeshes, I wanted it working in blender, so I can import, convert, then export it again.

In the future, when I crack SKEL, I will not use OBJ anyways, and this will be a moot point aside from an alternative model format.