Added option to force MTL filename's to be valid. Fixed TexConv not using converted file.
I forgot TexConv auto-adds the extension of the new file; great for everycase except this one. So I need to get the new_ext file somehow. I could do a scan and find the file with an ext. but I just hack it to a lookup table.

MTL's texture paths cannot contain spaces; which sucks. (They don't follow the double-quotes rule; I already checked). So if you want it to autograb your textures, you need to force them to be valid first.