Layout has been validated (cooraoborated may be better term)
A script to import WHM into 3dsMax confirms my thoughts that the layout is pos(XYZ ), quaternion(XYZW).

But the script also shows that it uses local transforms?

Skeleton was fairly close to being correct; and this avenue was a complete bust (unless I missed something)

I'll take the win of confirming my thoughts on the layout