Working Blender script
working-ish...

Bones work perfectly;
but I haven't figure out how to convert them properly to blender-space

So the bones can deform the mesh properly, but the bones aren't in the proper locations

Also some meshes lack boneweights but have skeletons? May be stored in excess? But bones is empty in mesh header?