Refactoring SGA
To reduce the tight coupling of 'Versioned' classes and their Unversioned base; the unversioned base no longer contains any version lookups.

Instead; the APIvX protocol should be used to get/specify types that each version is expected to have

This refactoring breaks alot of the old codebase; which expected most versioned types to 'figure it out' by being passed the version and the stream

The new APIvX format should ensure that
A) modules that want to be vX compatible define their 'version' of the vX classes (vX defines a helper to determine if a module has been properly defined as an API)
A.1) While modules can be apis, the preferred method is to use APIvX objects; while following the vX module syntax format for imports
B) 'Simple' to define their own arbitrary vX formats: Simply create a class with classvars and set the classes to their appropriate hook.
C) Registering versions is done in a single location; before each individual class needed to update a lookup table with the versioned classes & there was only one global table (per class), meaning swapping the version contexts was impossible. Now, we still have one global lookup table; but it's easier to pass a version lookup table and avoid worrying about ensuring other tables are up to date

This is kinda a rant, but this is a massive API change
With tests not fully implemented for SGA, alot of functionality is broken; and the tests created need to be refactored to account for the changes