New API for SGA (v2/v5/v7/v9)
It still really sucks; but to avoid the problem that the OFFICIAL RELIC ARCHIVE TOOL has; each `version` reimplements it's own codebase

I get around this as much as I can with ABCs but it is ugly.

Ranting about official Archive Viewer; admittedly, they are distributed with CoH2 (and DowIII) so as long as it works for those formats; great! But then why bother supporting past formats?

Dow II (v5) is among the formats which is supported (v4 is the earliest supported according to their version-assertion logic; COH1 I assume) but the checks to support that format don't seem to line up with my own research

I'd have assumed this was because the switch to the Essence Engine (over whatever DOW1 uses) but COH2 and DOW2  both use the new Essence engine; so... ?

As an aside; that explains why DOW1's enumerations for FileStorageType are different.

I pray this is the last time I rewrite this API; I moved away from the clunky 'make no assumptions' strategy I had done before (which was very straightforward, but a lot of boilerplate) with a simple read/write at the archive level; with unpack/pack for helpers used for serializing data .

I need to get better at typing (figuring out generics of some kind) to make ArchiveABC  auto-type things like Folder/File/Drive definitions and their respective pack-aware helpers.

We also stop using those nifty ptr objects; as much as I liked the idea of 'slicing' a binary stream; it's annoying to debug, and only hides where a stream is pointing:

perhaps using a slice which only warns/errors when reading/writing outside the bounds (allowing seeks); and will only read to the 'end of the slice'